Subs 1.0.0.0 -> 198.18.1.[1|2] in network checker
It turned out that some network equipment filters
out those fake addresses 1.0.0.0 which we used for
network checker probing packages. 198.18.1.0/24 is
one of those ranges which are reserved for benchmarking
network connectivities http://tools.ietf.org/html/rfc2544,
so it seems to be appropriate to use
those addresses for this case by default.

Change-Id: Iddb7e71de8746c5e55de5468fadc0b6d45c93800
Closes-Bug: #1291655