Prepare `network-checker` package
Prepare `network-checker` package to upload it to `pypi`. It is
required for `fuelmenu`. So it will be added to the `requirements.txt`
when it will be available on `pypi`.

Move content of `setup.py` file to `setup.cfg` file.

Add `Makefile` for quick and handy `release` command.

Some changes (files) were taken from `openstack-dev/cookiecutter.git`.

Change-Id: I8213b6214a20b27c0f7162b696690e584ab37ad7
Partial-Bug: #1502047