Prepare `network-checker` package
Prepare `network-checker` package to upload it to `pypi`. It is
required for `fuelmenu`. So it will be added to the `requirements.txt`
when it will be available on `pypi`.

Move content of `setup.py` file to `setup.cfg` file.

Add `Makefile` for quick and handy `release` command.

Some changes (files) were taken from `openstack-dev/cookiecutter.git`.

Change-Id: Ibf884a5eeb006d26d49190e99efadb67c5681984
Partial-Bug: #1502047