Prepare `network-checker` package
Prepare `network-checker` package to upload it to `pypi`. It is
required for `fuelmenu`. So it will be added to the `requirements.txt`
when it will be available on `pypi`.

Move content of `setup.py` file to `setup.cfg` file.

Add `Makefile` for quick and handy `release` command.

Some changes (files) were taken from `openstack-dev/cookiecutter.git`.

Change-Id: Id5b2cb606dc8e5bdce5aba845432a5bd0a1ee419
Partial-Bug: #1502047